[#57] Delete empty parents when deleting last file
After deleting a file (aka resource in FileSystemStorage) we now bubble
up and delete empty directories.

Keep in mind that this may introduces a race conditon when 'A' deletes
last file of a directory and 'B' in same time PUTs a file to this same
directory. Didn't handle that because I was not able to reproduce such
scenario on my machine.